requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
adrf==0.1.4
    # via -r requirements.in
asgiref==3.7.2
    # via django
async-property==0.2.2
    # via adrf
async-timeout==4.0.3
    # via redis
dj-database-url==2.1.0
    # via -r requirements.in
django==5.0.2
    # via
    #   -r requirements.in
    #   adrf
    #   dj-database-url
    #   djangorestframework
djangorestframework==3.14.0
    # via adrf
gunicorn==21.2.0
    # via -r requirements.in
packaging==23.2
    # via gunicorn
psycopg[binary]==3.1.18
    # via -r requirements.in
psycopg-binary==3.1.18
    # via psycopg
pytz==2024.1
    # via djangorestframework
redis==5.0.1
    # via -r requirements.in
ruff==0.2.2
    # via -r requirements.in
sqlparse==0.4.4
    # via django
typing-extensions==4.9.0
    # via
    #   dj-database-url
    #   psycopg
